Captured CMD-S event
I got tired of accidentally pressing CMD-S when trying to save something in Joomla. This little bit of code will capture the CMD-S (or CTRL-S) and save the form (not the browser page) any time you're editing an entity (module, article, menu item, etc...). It tested and it works on Mac and Windows on all major browsers. Keep it if you like it; otherwise, you won't hurt my feelings.